requirements.txt
streamlit>=1.33
pandas>=2.2
numpy>=1.26
scikit-learn>=1.4
tensorflow>=2.13
joblib>=1.3
pillow>=10.0
pyarrow>=15.0
